requirements.txt
flask
pywebview
playwright
requests
pystray
Pillow
